Change Components name and clean some code
