--- 42-transcendence/client/src/components/Layout.tsx
@@ -1,15 +1,0 @@
-import React, { ReactNode } from 'react';
-import NavBar from './home/NavBar';
-import { motion } from 'framer-motion';
-interface Props {
-    children: ReactNode;
-  }
-  
-function NavBarLayout({ children }: Props) {
-    return (<div className="content">
-        <NavBar />
-        {children}
-    </div> );
-}
-
-export default NavBarLayout;
